AtCoder/ABC200/C/main.go
package main

import (
	"bufio"
	"fmt"
	"os"
	"strconv"
)

var sc = bufio.NewScanner(os.Stdin)

func readInt() int {
	sc.Scan()
	i, _ := strconv.Atoi(sc.Text())
	return i
}

func run(n int, as []int) int {
	dp := make([][]int, 200)
	ans := 0
	for _, a := range as {
		idx := a % 200
		ans += len(dp[idx])
		dp[idx] = append(dp[idx], a)
	}
	return ans
}

func main() {
	sc.Split(bufio.ScanWords)
	n := readInt()
	as := make([]int, n)
	for i := range as {
		as[i] = readInt()
	}
	fmt.Println(run(n, as))
}
